Tests moved to separate files + other stuff
New math stuff and more utility functions. Also moved all tests to
separate files to make it simpler to compile the library without
containing the tests in the binary. Having tests next to the functions
was a bit problematic to compile.

Also the makefile should be way simpler to maintain now