Add function to "optimize" polygons
This function will remove points in a polygon that aren't necessary when
rendering the polygon. This is done by looking for points that have a
point on both sides with X or Y axis being the same value.

This shouldn't be called if the polygon will be modified at some point,
since running this function will result in loss of details.