Prevent cases where texture size would be null
In some cases the entity might not have a texture, and in those cases
the texWidth and texHeight variables would be null. That didn't trigger
the "Tried to render an entity with a texture with size of <= 0" error,
but could have caused memory errors in some conditions.

The fix is to just skip texture size checks if the entity doesn't have a
texture on it and just continue with the texture size of 0