Start building a simple level editor for 2D maps
So far it has a basic reference grid that can be moved around with the
middle mouse button and resized with the scroll wheel. The grid position
can also be reset with the 'home' key.

The goal for this level editor is to help with making tile based simple levels for 2D games so that hard coding level data into the game code could be avoided. This could also open the door for game modding or something, who knows ;)