Add a scene specific AddObject function
The AddScene() version of AddObject() should be used when adding Scenes
to other Scenes as children. AddObject() *can* still be used, but
AddScene() should be preferred.

The AddScene() function implements a recursive check to figure out if
the function being added as a child is actually a parent of itself. This
will lead to a stack overflow otherwise

An example situation:
```
Scene sceneA;
Scene sceneB;

sceneA.AddObject(&sceneB);
sceneB.AddObject(&sceneA);
```
This code would lead to a segfault due to a stack overflow. A simple
check for adding a scene into itself can be done by comparing the
SceneObject pointers, but any further checks can't be done, because the
SceneObjects aren't aware of their parents. This is worked around with
AddScene() by keeping track of parent Scenes and if everything is fine,
the Scenes are added as children like any other SceneObject would be