Destroy the Text entity sprite when moving or running out of scope
When constructing a new text entity by using an already declared one,
there was a possibility that the old sprite would be left hanging and
never destroyed.

Also when a text component runs out of scope, there's no need to worry
about someone else using the sprite it has, because the text component
has created the sprite by itself. Thus it can be safely destroyed
freeing the memory