Start working on a way to construct entities late
In some cases the entity cannot be constructed during declaration, so it
would be really useful to be able to construct them later on during the
games lifespan. Move and copy constructors are a pain to deal with and a
quick way to end up with pointer issues. Also letting the object
construct itself saves a little bit of performance