Improve the main menu quit button
It is now possible to know if the user clicked the "Quit" button in the
menu. This is useful, because you don't want to load in the game assets
if the user had decided to quit the game.

To use this "feature", you'll want to get the mainmenu.Launch() return
value and check if its false. If its false, you can set the
application_running variable under the game class to false and skip the
rest of the start() function